Fix: bi_reverse pure shim binding uses u8 for len (matches spec)
The spec says (code: u32, len: u8) -> u32. My binding had len: i32,
making the generated test pass i32 where the fn expects u8 — compile
fail. The C shim signature is (unsigned, int) but ctypes coerces the
Python value correctly; only the Rust literal type needs to match the
Rust fn signature.

+++ alchemist/extractor/c_shim_fuzz.py
@@ -298,7 +298,7 @@ ZLIB_SHIM_PURE_BINDINGS: dict[str, CShimPureBinding] = {
         name="bi_reverse",
         args=[
             StateFieldSpec("code", "u32", fuzz_u32),
-            StateFieldSpec("len", "i32", lambda rng: rng.randint(1, 15)),
+            StateFieldSpec("len", "u8", lambda rng: rng.randint(1, 15)),
         ],
         argtypes=[ctypes.c_uint, ctypes.c_int],
         restype=ctypes.c_uint,
